+++ drools-persistence-jpa/src/main/java/org/drools/persistence/PersistenceContext.java
@@ -7,7 +7,7 @@ public interface PersistenceContext {
 
     void persist(SessionInfo entity);
 
-    public SessionInfo findSessionInfo(Long id);
+    public SessionInfo findSessionInfo(Integer id);
 
     boolean isOpen();
